+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2024-10-17
+
+### Added
+
+- Support for health-check-interval configuration options from integration with Content Cache Backends Config charm.
+- Support for backends-path configuration options from integration with Content Cache Backends Config charm.
+- Support for proxy-cache-valid configuration options from integration with Content Cache Backends Config charm.
+
 ## 2024-10-07
 
 ### Added